+++ resources/views/layouts/partials/backtotop.blade.php
@@ -1,0 +1,3 @@
+<a class="backtotop js-backtotop" title="{{ trans('general.back-to-top') }}" href="{{ Request::url() }}#top">
+    <span class="backtotop__icon icon--arrowUp"></span>
+</a>